2022.10.30(DM) edit friend management's files and move another folder
